Alternate fair/unfair lock modes in the connection pool to improve throughput
Acquiring locks unfairly in `OpenConcurrencyLimiter.releasePermit`
revealed a bug in `OpenConcurrencyLimiter.acquirePermitOrGetAvailableOpenedConnection`,
which manifested itself by failing the `pool-checkout-minPoolSize-connection-maxConnecting.json`
test. A fix had to be implemented together with the performance-related change.

Another tiny change is moving `pool.release(openConnection)` out of the guarded section
in the `OpenConcurrencyLimiter.tryHandOverOrRelease`: it does not have to be there,
and when not only fair locking is used, this change has a significant positive measurable effect
on the connection cycle (check out/check in) throughput.

JAVA-4435