Address review concerns
1) The type hierarchy were simplified where possible. For example, `ValueBoostSearchScore`/`PathBoostSearchScore` are now
both just `SearchScore`, and there is no more `BoostSearchScore` type (also, there was a bug in that hierarchy,
but it was not the reason for removing `BoostSearchScore`). The same thing was done to `near` and `range` search operators.
If we need to have any of those types in the future, we can always add those types back, but exposing them now without the need seems unwise. We may need such intermediate types, for example, when we need to add a method common to all `boost` search score modifiers such that it is not present on all `SearchScore` types. The only intermediate
specific type what was not removed is `CompoundSearchOperator`. It is there because all its subtypes need to have the `score` method returning `CompoundSearchOperator` (as opposed to returning `SearchOperator`),
and `CompoundSearchOperator` is a place where such a method may be declared.

2) I clarified how we extract data from `Instant`/`Date`, because we do it in a lossy way.

3) The order of `SearchOperator.near` parameters was changed.

4) I left the `gteLte` approach when specifying intervals for the `range` operator intact.
This is mostly a matter of taste to me, but if someone points out somewhat objective arguments against the approach,
I will be open to reconsider.

JAVA-4190