Add missing condition for whether to sleep in server monitor
There is a missing check in an already-complicated conditinal statement
for whether to sleep in the server monitor before executing the next check.

The condition is included in the server discovery and monitoring specification
pseudo-code but missing in the driver, and there are currently no tests for it.

This commits adds the missing condition to the conditional statement.  Effectively,
this will reduce the number of sockets that the monitor opens on a server that is
in the process of shutting down but still accepting connections.

JAVA-4730