Update MongoQueryException to extend MongoCommandException
The MongoQueryException class currently extends MongoServerException,
the reason being that prior to MongoDB 3.2 the find helper was implemented
using the OP_QUERY wire protocol message, which is not a command.  But now
that 3.6 is the minimum required server version, the find helper is always
implemented via the find command.  So MongoQueryException can now extend
MongoCommandException and include the full command response document.

JAVA-4676