Initial driver-kotlin-sync implementation
Wraps the driver-sync to provide a kotlin sync API

In general a faithful wrapping of the sync java driver. Uses reified
overloads to provide synaptic sugar instead of passing resultClass.

MongoIterable deviates from the sync API so to not implement Iterable in
order to protect the user from leaving open cursors. In Kotlin there are two Iterable
style implementations: Iterable and Sequence neither are auto closable.
Sequence is more appropriate here as it lazily consumes the iterator,
but it does not auto close. See: https://youtrack.jetbrains.com/issue/KT-34719
So MongoIterable provides a loan method that treats the cursor as a
Sequence and automatically closes after use.

JAVA-4869