Remove dead code in ReplyHeader
The dead code is a remnant of when the driver supported the
full range of OP_REPLY wire protocol message usage.  Since the driver
now only runs against MongoDB releases that support OP_MSG,
OP_REPLY is only used for the response to the `hello` command
in the handshake, and therefore most of the OP_REPLY-handling code is
no longer on any execution paths.

JAVA-5204