Add support for the unified test schema version 1.17
Note that despite this commit marking 1.16 as supported,
it actually is not. The existing practice
is to mark as supported all the versions up to the one that is
either fully or partially supported, despite them
being likely not actually supported. When someone discovers
that a feature is used in a test, but is not implemented,
then he implements that feature.

JAVA-4936