Convert Bson to BsonDocument for hint
Unlike every other instance of Bson in the CRUD API, the hint option for write
operations was not converted to a BsonDocument prior to passing it down to the
operation layer.  This is a problem because the operation layer doesn't have a
CodecRegistry available, and so can not properly due the conversion.

In this commit, all hint options are converted to BsonDocument in the same place
as all the other Bson options, where a CodecRegistry is available to properly
do the conversion.

JAVA-5357