Move shared library resource entries to :graalvm-native-image-app
Having these entries results in a larger native image,
even if the shared libraries are not used by
the application. Therefore, we won't supply
the entries with our JARs, but will document
this for users and propose the `resource-config.json`
in :graalvm-native-image-app as an example
where users may pick the necessary entries from
if need be.

JAVA-5219