Add JDK-specific reachability metadata to `org.mongodb:mongodb-driver-core`
Without this metadata an application using
the driver won't work, and there is no other
place this metadata may come from, unless
it is gathered via the GraalVM tracing agent.

It's weird that we have to include reachability
metadata specific to JDK, but there is
nothing to do about it, other than saying that
users are to collect the metadata for their
applications themselves. The latter actually
seems like the only right way to do this, but it
conflicts with the fact that there is Oracle's
reachability metadata repository
(https://github.com/oracle/graalvm-reachability-metadata).

JAVA-5219