Add artifact tracing on publish (#1364)
This commit adds papertrail tracing on release to all modules we release.
This tracing will run on every snapshot and release.

The change to publish.gradle was done because we need to ensure that
the generated .pom, .md5, and .sha1 files are available on disk for
tracing. We will now also "publish" these files to a location on disk
that enables simple tracing in Papertrail. The filename patterns were
chosen because the gradle publish command also includes non-release
files with duplicate names that we do not want to trace.

JAVA-5415