Replace lazy command document with non-lazy one
Justification: this is the only operation that uses a lazy document for the command.  It was only lazy in the first place because of the lack of splitting in the initial implementation, but now that there is splitting there is no longer a need, and the fact
that it's an outlier would make it confusing for future readers.

JAVA-5529

Co-authored-by: Jeff Yemin <jeff.yemin@mongodb.com>