Improve selection criteria for srvMaxHosts (#1590)
Improve selection criteria for srvMaxHosts

This PR improves the selection criteria for srvMaxHosts. The net result is that it will keep existing nodes, discard deleted nodes, and fill up to srvMaxHosts from a random shuffle of newly discovered nodes from the SRV response. It also updates the prose test to match the spec.

JAVA-4400