Ensure custom KProperty include the name in the hashcode
The `pathCache` utilised the string representation of the KProperty instance.
The custom implementations didn't include the calculated path value, this can
lead to naming collisions in the pathCache key.

This commit adds `hashCode` and `equals` methods to ensure the `name` value is included
in our custom implementations of `KProperty1`. Finally, the cache key is explicitly
created using the same default `Any.toString()` implementation to ensure that all
implementations of `KProperty1` use the same logic to create the cacheKey.

JAVA-5868