Ensure custom KProperty include the name in the hashcode (#1731)
The `pathCache` utilised the string representation of the KProperty instance.
The custom implementations didn't include the calculated `path` value, this can
lead to naming collisions in the `pathCache` key.

This commit adds `hashCode` and `equals` methods to ensure the `name` value is included
in our custom implementations of `KProperty1`. Finally, the cache key now uses the `hashCode`
for the cacheKey.

JAVA-5868
Co-authored-by: Ross Lawley <ross@mongodb.com>