Added nexus-publish plugin (#1751)
This plugin will eventually be used for the migration to the central portal
Currently still uses the legacy osshr sonatype location.

As the nexus-publish plugin can only be configured in the root project
the username and password logic was moved out of conventions/publishing.gradle.kts
and into the root buid.gradle.kts.

Note: The nexus-publish plugin builds upon the gradle maven-publish
plugin so even though the configuration is in the root project,
it will only publish projects that use the publishing convention.

JAVA-5899